Novel/Episode table에 score 칼럼 추가, Novel table에 hit 칼럼 추가

+++ src/table/Comment.ts
@@ -31,7 +31,7 @@ export let table = {
     allowNull: false
   },
   score: {
-    type: Sequelize.INTEGER(5),
+    type: Sequelize.INTEGER(1),
     allowNull: false
   }
 }
